requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
protobuf==3.15.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
yfinance==0.1.59
